produtct엔티티에 이미지  정보 등록
- DB 등록은 완료
- 테스트 케이스 다시 살펴봐야함

+++ src/main/java/org/shopmoon/domain/AttachImageVO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class AttachImageVO {
 
 	private String uploadPath;
-	private String uuid;
+	private String uuid; // PK
 	private String fileName;
-	private int productNo;
+	private Long productNo; // FK
 }
